+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # @project-version@ (@build-time@)
 
-* Added missing default settings for target style for soft friend and interact target.
-* Fixed a bug that prevented the scaling and transparency from being applied to mousover units [GH-519].
+* Fixed a Lua error with Auras widget the occurred when a Priest uses Mindcontrol on another player [GH-520].